removed unused imports in views and tried to get rid of a few warnings

+++ Management/views.py
@@ -1,12 +1,12 @@
 from django.db.models import Q
-from django.http import HttpResponse, HttpResponseRedirect
-from django.shortcuts import render, redirect, get_object_or_404
+from django.shortcuts import render, redirect
 from django.views import View
 from django.views.generic import ListView
 from .email import Email
 from .models import User, Section, Course, UsersToCourse
 from .forms import SectionForm, CourseForm, UserForm, UserToFrom, CourseEditForm, SectionEditForm
 from django.contrib.auth import logout
+
 
 # Create your views here.
 # Need to create a landing page after login
@@ -26,7 +26,7 @@ class Home(View):
             password = request.POST['password']
             m = User.objects.get(email=request.POST['email'])
             badPassword = (m.password != request.POST['password'])
-        except:
+        except User.DoesNotExist:
             noSuchUser = True
         if user == '' and password == '':
             return render(request, "main/home.html", {"message": "Email and password required."})
@@ -74,7 +74,7 @@ class MainHome(View):
 class courses(View):
     def get(self, request):
         if 'messageC' not in request.session:
-            request.session['messageC']=""
+            request.session['messageC'] = ""
 
         msg = request.session['messageC']
         request.session['messageC'] = ""
@@ -97,10 +97,11 @@ class courseAdd(View):
         if Course.formAdd(form):
             request.session['messageC'] = "Course Successfully added!"
             courses_view = courses()
-            #return render(request, 'main/Course/courseAdd.html', {'form': form, 'message': "Course Successfully added!"})
+            # return render(request, 'main/Course/courseAdd.html', {'form': form, 'message': "Course Successfully
+            # added!"})
             return courses_view.get(request)
         else:
-             form = CourseForm()
+            form = CourseForm()
         return render(request, 'main/Course/courseAdd.html', {'form': form,
                                                               'message': "Course ID Already Exists."})
 
@@ -115,7 +116,6 @@ class courseEdit(View):
 
         return render(request, "main/Course/courseEdit.html", context)
 
-
     def post(self, request, course_id):
 
         course = Course.getCourse(course_id)
@@ -123,7 +123,8 @@ class courseEdit(View):
         if course.formSave(form):
             request.session['messageC'] = "Course was successfully edited."
 
-            #return render(request, 'main/Course/courseEdit.html', {'form': form, 'message': "Course was successfully edited!"})
+            # return render(request, 'main/Course/courseEdit.html', {'form': form, 'message': "Course was
+            # successfully edited!"})
             courses_view = courses()
             return courses_view.get(request)
         else:
@@ -149,7 +150,7 @@ class courseDelete(View):
 class usersInCourse(View):
     def get(self, request, course_id):
         if 'messageUIC' not in request.session:
-            request.session['messageUIC']=""
+            request.session['messageUIC'] = ""
         msg = request.session['messageUIC']
         request.session['messageUIC'] = ""
 
@@ -196,14 +197,15 @@ class userToCourseDelete(View):
         user = UsersToCourse.objects.filter(assignment=email_id, courseID=course_id)
         try:
             for x in user:
+                if not isinstance(x, UsersToCourse):
+                    raise TypeError("x is not an instance of UsersToCourse")
                 x.delete()
-        except:
-            not isinstance(user, UsersToCourse)
-        request.session['messageUIC'] = "User successfully deleted."
+        except TypeError as e:
+            request.session['messageUIC'] = str(e)
 
         # Redirect to a success page or back to the list of courses
 
-        #return redirect("usersInCourse", course_id=course_id)
+        # return redirect("usersInCourse", course_id=course_id)
         user_to_course_view = usersInCourse()
         return user_to_course_view.get(request, course_id)
 
@@ -232,7 +234,7 @@ class sectionAdd(View):
 
         if 'messageS_invalid_form' in request.session:
             context = {'form': form, 'course_id': course_id, 'people': user,
-                                                                'message': request.session['messageS_invalid_form']}
+                       'message': request.session['messageS_invalid_form']}
             return render(request, 'main/Section/addSection.html', context)
 
         return render(request, 'main/Section/addSection.html', {'form': form, 'course_id': course_id, 'people': user,
@@ -244,7 +246,7 @@ class sectionAdd(View):
         if Section.formAdd(form, course_id):
             request.session['messageS'] = "Section successfully added."
 
-            #return redirect('sections', course_id=course_id)
+            # return redirect('sections', course_id=course_id)
             sections_view = sections()
             return sections_view.get(request, course_id)
         else:
@@ -254,9 +256,6 @@ class sectionAdd(View):
         request.session['messageS_invalid_form'] = "Section ID Already Exists."
 
         return sectionsAdd_view.get(request, course_id)
-
-
-
 
 
 class sectionEdit(View):
@@ -388,4 +387,3 @@ class userDelete(View):
 def userLogout(request):
     logout(request)
     return redirect('login')
-
